requirements.txt
fastapi==0.134.0
uvicorn[standard]==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1

psycopg2-binary==2.9.10

apscheduler==3.11.2

httpx==0.28.1

python-dotenv==1.0.1
loguru==0.7.3
